// packages/back/src/forms/types.ts
export interface FullForm extends Form {
  temporaryStorageDetail: TemporaryStorageDetail;
  transportSegments: TransportSegment[];
}